# fix(util): improve performance of classToInvokable