--- a/lib/utils/classToInvokable.js
+++ b/lib/utils/classToInvokable.js
@@ -10,6 +10,12 @@ function classToInvokable(Class) {
   return new Proxy(Class, {
     apply(Target, thisArg, args) {
       return new Target(...args);
+    },
+    construct(Target, args) {
+      return new Target(...args);
+    },
+    get(target, p) {
+      return target[p];
     }
   });
 }