### Pull Request check-list

_Please make sure to review and check all of these items:_

- [x] Does `npm run test` or `npm run test-DIALECT` pass with this change (including linting)?
- [x] Does the description below contain a link to an existing issue (Closes #[issue]) or a description of the issue you are solving?
- [x] Have you added new tests to prevent regressions?
- [x] Is a documentation update included (if this change modifies existing APIs, or introduces new ones)?
- [x] Did you follow the commit message conventions explained in [CONTRIBUTING.md](https://github.com/sequelize/sequelize/blob/master/CONTRIBUTING.md)?

### Description of change

Script:

```js
'use strict';

class A {
  constructor(a, b) {
    this.rng = Math.random();
    this.a = a;
    this.b = b;
  }
}

A.a = 1;
A.b = 2;

function classToInvokable(Class) {
  return new Proxy(Class, {
    apply(Target, thisArg, args) {
      return new Target(...args);
    },
    // NEW
    construct(Target, args) {
      return new Target(...args);
    },
    get(target, p) {
      return target[p];
    }
  });
}

const B = classToInvokable(A);

console.log(new A(1, 2), new B, new B(1, 2), B(1, 2));

console.time('class');
for (let i = 0;i<0xFFFFF;++i) {
  const a = new A();
}
console.timeEnd('class');


console.time('proxy new');
for (let i = 0;i<0xFFFFF;++i) {
  const b = new B();
}
console.timeEnd('proxy new');

console.time('proxy invoke');
for (let i = 0;i<0xFFFFF;++i) {
  const b = B();
}
console.timeEnd('proxy invoke');

const a = new A();
console.time('class');
for (let i = 0;i<0xFFFFF;++i) {
  A.a + A.b;
}
console.timeEnd('class');


const b = new B();
console.time('proxy');
for (let i = 0;i<0xFFFFF;++i) {
  B.a + B.b;
}
console.timeEnd('proxy');
```
```
A { rng: 0.583534638354642, a: 1, b: 2 } A { rng: 0.6534978634885156, a: undefined, b: undefined } A { rng: 0.9860787273348626, a: 1, b: 2 } A { rng: 0.9952360726330047, a: 1, b: 2 }
class: 17.758ms
proxy new: 2372.765ms
proxy invoke: 92.412ms
class: 2.418ms
proxy: 119.242ms

 === WITH TRAPS ===

➜  sequelize git:(chore/performance) ✗ node perf.js
A { rng: 0.7856633966445696, a: 1, b: 2 } A { rng: 0.9549997856154535, a: undefined, b: undefined } A { rng: 0.14584827492246366, a: 1, b: 2 } A { rng: 0.9905286808234037, a: 1, b: 2 }
class: 17.300ms
proxy new: 63.847ms
proxy invoke: 60.784ms
class: 2.299ms
proxy: 76.098ms
```

As you can see the Proxy is dramatically slower without the traps.